+++ gems/shaolin-cli/lib/shaolin/cli.rb
@@ -2,6 +2,7 @@ require_relative "cli/version"
 require_relative "cli/naming"
 require_relative "cli/generators/module_generator"
 require_relative "cli/generators/new_app_generator"
+require_relative "cli/generators/field_generator"
 
 module Shaolin
   module CLI